+++ Tracker-ios/Supporting Files/Helpers/UICreator.swift
@@ -70,7 +70,7 @@ struct UICreator {
         return textField
     }
 
-    func makeSearchTextField() -> UISearchTextField {
+    func makeSearchTextField(withTarget action: Selector? = nil) -> UISearchTextField {
         let searchField = UISearchTextField()
         searchField.placeholder = "SEARCH".localized
         searchField.backgroundColor = .ypGrayLight.withAlphaComponent(0.12)
@@ -83,6 +83,9 @@ struct UICreator {
             leftView.tintColor = UIColor.ypGray
         }
         searchField.clearButtonMode = .never
+        if let action {
+            searchField.addTarget(nil, action: action, for: .editingChanged)
+        }
         return searchField
     }
